// src/api_main/models/UserLoginRequestModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BaseWebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserLoginRequestModel
 */
export interface UserLoginRequestModel {
    /**
     * 
     * @type {string}
     * @memberof UserLoginRequestModel
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof UserLoginRequestModel
     */
    password?: string;
}

/**
 * Check if a given object implements the UserLoginRequestModel interface.
 */
export function instanceOfUserLoginRequestModel(value: object): value is UserLoginRequestModel {
    return true;
}

export function UserLoginRequestModelFromJSON(json: any): UserLoginRequestModel {
    return UserLoginRequestModelFromJSONTyped(json, false);
}

export function UserLoginRequestModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserLoginRequestModel {
    if (json == null) {
        return json;
    }
    return {
        
        'username': json['username'] == null ? undefined : json['username'],
        'password': json['password'] == null ? undefined : json['password'],
    };
}

export function UserLoginRequestModelToJSON(json: any): UserLoginRequestModel {
    return UserLoginRequestModelToJSONTyped(json, false);
}

export function UserLoginRequestModelToJSONTyped(value?: UserLoginRequestModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'username': value['username'],
        'password': value['password'],
    };
}